+++ skxray/core/utils.py
